requirements.txt
python-dotenv==0.7.1
psycopg2==2.7.3.2
Flask==0.11.1
Jinja2==2.8
MarkupSafe==0.23
Werkzeug==0.11.10
click==6.6
gunicorn==19.6.0
itsdangerous==0.24
requests==2.10.0
wsgiref==0.1.2
flask-sqlalchemy
